Some heavy optimizations
One optimization flatly increases speed 4x.
The remaining ones direct the search so it finds much smaller solution
in much shorter time.
It is still not enough to pass the disabled tests but I'm pretty proud
of that regardless.
Well, it might be enough. I might overthinking this and maybe the
algorithm doesn't need multiple passes at all and it can find the best
solution in the first go. I must to think of some proof. Sadly, I've
already wasted a lot of time to create a nice progress bar.