Extracted prime implicant code as a separate class
The code is already cleaner even that half of it still uses the name
"minterm" instead of "prime implicant".
It's annoying that tests changed. I fixed every bug that I could find
and now it seems like the results or just different but not incorrect.
This is probably some change in the sorting order at some stage, which
screws with the heuristic later. Actually, now the algorithm seem to
prefer inputs from the beginning of the list, which was always the
intention, so maybe that's an improvement.