Optimizing via brute force instead of a heuristic
This is not entirely finished but it took a lot of time and I need a
backup before continuing.
I'm not even sure if this is an NP problem. It kinda looks similar to NP
but I have trouble finding an example for which a heuristic gives worse
result. (Maybe it because more complicated examples can be run only with
use of heuristic to not take ridiculously long time so I don't have a
comparison.) I'm gonna need to research that more thoroughly.