Added a proper `README.md`
OK, technically there were a `README.md` before that so is was not added
per se. However, the old version was so short and useless that it
doesn't really count. It was just something added by GitHub in the
initial commit.
Now it has everything it should except the things I forgot.