Moved creation of `Progress` like 4 functions down
There is no need now to pass object of `Progress` into function after
function. (Although, now the name of the function is passed instead so
it comes up kinds same. At least a few instances of `Progress` do not
exist at the same time, so it is cleaner in that way. Hmmm... actually
that may improve accuracy of some progress bars.)