Moved some repeating code to the class `Progress`
Sometimes progress has to perform some complicated calculations tailored
to a specific operation to measure but it is common that it is just
counting linear increases. This code is basically the same in every
place so it was extracted into a common class.
There was also a few bugs fixed.