Added C++ output format
I though VHDL was complicated but C++ definitively wins the prize for
the most corner cases.
The generated code is a stateless functor which additionally has static
functions that do the same. I've decided not to use plain functions
because it was easier to put definitions of input/output types in the
same scope if everything was in a class. It can also be useful for
someone that it is a functor?