Switched to C++20
This gives a nice 3% speed boost on GCC (and I need all the speed I can
get here) but primarly this will allow usage of `[[likely]]` and
`[[unlikely]]` attributes which may help a lot in some places.
Also, changed how UTF-8 characters are handled in MSVC because `u8`
literals are a menace and I won't deal with them.