Pre-reserving memory based on an estimation
The estimation is based on file size, assuming the numbers are uniformly
distributed over the allowed range, with 10% margin of error.
This may decrease time in average case about 1 or 2 percent and may
reduce memory requirements during loading. (Although, the memory may not
be such a problem before the later stages.)