Refactored the class `Implicant`
This was mostly done to make future optimizations easier to write but it
also proved to speed up the execution significantly by itself.
Changing the way the data is stored from "ones" & "zeros" to "bits" &
"mask" speed up finding prime implicants 2 times.
Not storing bit count but calculating it every time, reduces memory when
finding prime implicants by 30% (and also seem to speed up the program
because this value is used not that often and keeping it updates was
expensive).
There is also a bunch of other changes and optimizations that gave much
smaller gains.