Added heuristic for finding prime implicants
The code is more complicated now but at least it is much slower.
The heuristic is pretty bad. It can find implicants for about 10% of the
minterms and they are not the prime implicants. The main algorithm
cannot use the most important optimizations to use the data from
heuristic so it is slower on the big O level.
The idea is that a good heuristic produces some decent approximation of
all prime implicants and then the main algorithm takes over, fixing it,
refining, and filling the gaps. This commit is probably a step in the
right direction for that.