Rewritten delay of progress bar
The progress bar shows now after 3 seconds, not 1. However, it is now
time from the start of the application and not from the creation of the
`Progress` object, so there will no more be time gaps between progress
bars.
Also added a function to print this time, for debugging purposes.