A cool and new heuristic
I've kinda given up on the previous heuristic. I'm not sure it is
salvageable. It was fast but really bad.
The new one uses a divide-and-conquer approach. A single iteration of it
is about 50% slower than the old heuristic but it gives better results
and it can be repeated to find more cubes. (Is this how we are supposed
to call implicants?) This is promising approach that may be even viable
to be used with 32-bit input. However, I'm starting to lose hope that
I'll be able to get all prime implicants by refining heuristic with
brute-force. This just takes too long. Maybe a partial refinement...