A few failed optimizations
There doesn't seem to be any noticeable difference in performance. Most
of tries (not included in this commit) did actually worsened the speed.
Either the C++ optimizer is smarter than me or the code is bottlenecked
by the cache locality. (Tricks I tried to improve it didn't work.)

These changes may or may not make the optimization more consistent in
different compilers. There is also some code cleanup.