Two cosmetic improvements of the heuristic
These two changes don't really change anything. They are all the ideas
that worked during a whole day of trying to improve accuracy of the
heuristic. Every serious try to get a better estimation of where to cut
the minterms turned out to decrease both the speed and accuracy
significantly. Idn. maybe the simplest ideas are the best in such cases.
I think I'll leave this algorithm as is and focus on the things around
it.