Limited number of iterations of the heristics
The gains from subsequent iterations drop very quickly. Now, it stops as
soon as the current iteration fails to eliminate 50% of the remaining
minterms (which happens around the 2nd run). This seems pretty harsh but
benchmark has shown that it reduces the runtime by about 70% while the
result is not much worse than when it it run to the end. (Only the
number of smallest prime implicants is visibly worse.)