Simplify the heuristic even more
I made the heuristic painfully simple, which makes it over 25% faster
without changing the result. It seems that the simple it gets, the
better it works. Non of my elaborate algorithms (and I've wasted a lot
of time to devise them) seem to help. No, I'm not salty, you're salty.