Added an algorithm that expands implicants
Previously, first non-overlapping implicants were generated and then
then immediately a list of all prime implicants was generated out of
that. Now there is an additional step in the middle, that extends
implicant to overlap, so they are as big as possible (which also allows
merging more of them). This has the opposite effect from the last step
by reducing the number and size of implicants instead of blowing this
number up into cosmic values. It may be useful for bigger input counts.
Unfortunately, it is too slow for now.