Improved the algorithm that expands implicants
It was changed so much, it is practically a new algorithm at this point,
although, the core principle remains similar. It is A LOT faster now.
It's O(n) instead of O(n^2), where n is the amount of implicants. It
also seems to have better quality results but that may depend on chosen
criteria.
It's viable for the maximum input size now.