Excluded "don't cares" from bases for implicants
That furthers reduces the result size, the time and the memory usage,
depending on if and how many "don't cares" are present.
(By which I mean that the code that does that is written but currently
unused because it screws with the brute-force approach. It needs to wait
for command line options for heuristic.)