Additional iteration of finding implicants
This is kinda brute force-ish (but it doesn't really search all the
possibilities or anything near to that). It increases the time a few
times but it makes the result better and decreases the memory usage.
Technically, it can be repeated for even better result but the returns
are diminishing. Idn. if I'll implement the repeating thing.