Fixed progress bar freezing at the beginning
If the initial estimation of progress speed was too high, the progress
bar didn't show for extended period of time. To prevent that, for the
first 60 seconds progress is checked 2 times more often and for the
first 10 seconds 8 times more often. (There was some code in the past
that did something similar but more elaborate. It was removed because I
though it isn't needed. Tuns out it sometimes is but this simpler form
should also do the job.)