Fixed a few function qualifiers
`&&` was added to some functions to show that it cannot be called twice,
which is a pretty unconventional usage of it but I think it's better
than not marking it at all.
`const` was removed from a few functions to fix the optimization that
was supposed to remove some data early to decrease memory usage.