Fixed a serious bug that was revealed by MSVC
On MSVC the size of `std::uint_fast32_t` was 32 bits instead of 64 like
on GCC and Clang which triggered the bug. Now `std::uint_fast64_t` is
used instead. (The requirement for such bit variable is caused by a
weird order of operations there, that proved to be faster during tests.)