Added iterator to the class `Implicant`
It iterates over the minterms in order.
This reduces code duplication and it will be needed for the next commit.
Unfortunately, even that the new code could be theoretically equivalent
to the old one, the compiler optimizer didn't quite manages to notice
everything and the new program is about 2% slower.