Made removing redundant solutions actually fast
It went a few order of magnitudes down, landing on a pretty good speed.
The golden rule of optimization seems to be sort the data and then use
the fact that it is sorted.
(There still may be pathological kind of input data for which the new
approach is slow. I'll check before moving on to the next algorithm.)