Optimized reduction part of merging solutions
The "optimization" (if it can be called that) was just removing an
entire 300 line long class that was supposed to speed up the operation
and replacing it with some short code I copied from a few commits back.
This made the code about 50x faster. Don't do premature optimization
kids and if you do, don't try to be too fancy with it. Fancy and clever
optimizations seem to always never work.