Added an option to limit solution amount
At every step of the Petrick's Method it checks it there is too many
solutions and uses heuristic to reduce their number. (Currently the
heuristic is bad to non-existing - it just keeps first n solutions.)
By default the limit 256 solutions. (Which is a very bad limit - it
should be dependent on number of minterms or even something more obscure
internal value created by earlier steps.)
Still, it works well enough for now.

Also, the progress bar was simplified to not display the projected
number of final solutions. It was (questionably) useful only for high
input counts and they are not really viable without this heuristic,
which makes the projection obvious so there is no need to display it.