Optimized the case `--max-solutions=1`
In this case, we can use a much simpler method, which is about 6x faster
and still gives the same result (not strictly guaranteed, I don't think
but is seems to in practice and even if not, it'll be the same quality).
Also, added a few lines that eliminate solutions over `--max-solutions`
before the first merge is performed. This was strictly necessary for the
new method and for the old one it won't hurt and it's more consistent.