Moved `--help` & `--version` text to a new file
They were getting too long to be stored in the `main.cc` but this is old
news since the same change was already done in the (yet non merged)
development branch. The real reason of this change was exactly that.
Merging the `master` branch to the development branch were getting
annoying due to hard to fix conflicts so now `master` will mirror what
that branch did.