Generalized the new optimization for more cases
The new optimization can be used for all cases where the number of
solutions is limited (the option `--max-solutions`).
It would probably not work as well for unlimited solutions.
(It's 30x faster now! However, there is a risk of finding duplicated
solutions which wasn't a problem with the previous algorithm.)