Added heuristic `greedy` for merging solutions
It is extremely fast and it doesn't take much memory. The result is
worse than even the Petrick's limited to 1 solution but surprisingly not
that much worse. (Maybe the limiting needs more work, like a better sub-
-heuristic.)
Anyway, this is fast enough to handle the maximum of 32 inputs.
(And with ease; only a few seconds.)