Added output format `reduced-graph`
Saying that this output format has a lot of corner cases to handle is
not entirely accurate. It would be more precise to say that this output
format IS a lot of corner cases to handle. (It looks good, though.)
I can see a few more cases to test but I'm committing this now because
it works pretty well already and this seems like a good place to make a
checkpoint.