Added an utility class `First‘ and used it
It's a very common operation while printing data to do something
slightly different for the first item and creating a boolean flag every
time is repetitive and cumbersome. This class hides that behind a more
convenient interface. This is a very simple class but it helps a lot,
IMO and I can't believe it's not a common thing to do.