Decoupled options logic and definitions
The logic of parsing options (that could be reused in other programs)
was entirely decoupled from the options existing in this particular
program and moved to separate files.
(I really shouldn't make extensive changes to the code that I've already
significantly changes in another branch. The merge is going to be a lot
of pain. I'm even had to left some TOSOs of things that should be done
after that. A little masochistic on my side. Well, this is where the
inspiration have taken me and I'm willing to pay the price.)