Added heuristic `greedy` for the CSE
The method checks each pair of 2 values for common parts and takes the
first match. O(n^2) is not great and it can't handle the biggest inputs
but at least this approach is polynomial for a change. (It can still
handle significantly bigger inputs than the previous approaches.)

The code was retrieved from `bb7a79a95945be482a4d3927353a2c730b4835d4`.
It was an initial implementation of the CSE that was later replaced with
an exact method. It was almost entirely rewritten, though, to adjust it
to the new data structure and such.