Fixed shortcomings in handling optimization graph
In more complex cases, a lot of things were not removed from the graph
which should be removed, ending in bigger result than it could be.
(All 3 functions that clean up the graph had problems.)

Also, refactored the code to put things where they belong, which
drastically reduces the number of returned `std::pair`s.