Moved all the output related code to a new class
This way the mess is only in a single class (and this is a big mess).
Other classes are finally readable.
Beside moving the code, order of some functions was changed and some
changes were (of course) made to make the program work again after the
operation. All output functions were also changed to take `std::ostream`
instead of using `std::cout`.
Generally, though, the code remains relatively unchanged and it still
requires a lot of refactoring.