Reordered output formats to group prints better
In this order, printing in `OutputComposer` more often ends up with the
same strings next to each other in the list, which allows to use `NEXT`.
(The gain isn't is big as expected but writing a script that finds the
optimal order was fun - there are 4 such orders + their mirrored
versions, so ~1% of all possible orders are optimal. The one chosen here
is the most logical / similar to the old one.)

Also fixed a bug that caused "gate-costs" to take an incorrect execution
path (although, the result seemed to be correct).