Added a heuristic option `--rough-width`
This limits how many elements are used to build the graphs of subsets.
Only N elements that are occurring in sets most often are chosen.

It seems to finally put the heuristic on a speed that is acceptable for
the biggest examples, considering that I've just discovered that the
most time is spend on some step after this heuristic.
I didn't check how this affect the quality of results because, honestly,
I struggling to even make this work at all for 32 inputs.